+++ metrics-annotation/src/main/java/com/yammer/metrics/annotation/Gauge.java
@@ -20,8 +20,18 @@ import java.lang.annotation.Target;
  * annotated method's return value as its value.
  */
 @Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.METHOD)
+@Target({ElementType.METHOD, ElementType.FIELD})
 public @interface Gauge {
+    /**
+     * The gauge's group.
+     */
+    String group() default "";
+
+    /**
+     * The gauge's type.
+     */
+    String type() default "";
+
     /**
      * The gauge's name.
      */