+++ metrics-core/src/test/java/com/yammer/metrics/core/VMMFactory.java
@@ -1,5 +1,6 @@
 package com.yammer.metrics.core;
 
+import javax.management.MBeanServer;
 import java.lang.management.*;
 import java.util.List;
 
@@ -11,12 +12,14 @@ public class VMMFactory {
                                               OperatingSystemMXBean operatingSystemMXBean,
                                               ThreadMXBean threadMXBean,
                                               List<GarbageCollectorMXBean> garbageCollectorMXBeans,
-                                              RuntimeMXBean runtimeMXBean) {
+                                              RuntimeMXBean runtimeMXBean,
+                                              MBeanServer mBeanServer) {
         return new VirtualMachineMetrics(memoryMXBean,
                                          memoryPoolMXBeans,
                                          operatingSystemMXBean,
                                          threadMXBean,
                                          garbageCollectorMXBeans,
-                                         runtimeMXBean);
+                                         runtimeMXBean,
+                                         mBeanServer);
     }
 }